requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-multiselectfield==0.1.12
idna==3.4
PySimpleGUI==4.60.4
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
xlrd==2.0.1
